Update (see desc)
So it works, but I'm not sure if the Item Browser Potion category conditional is the proper way to check if an item is a potion. Also need to add support for recipes that have more than 14 ingredients, for now I just made a const int that should eventually be deleted but will make it a little easier to find all the places where it has to be changed.